Change Int to Long.

+++ api_generator/tests/references/kotlin/EntityWithPropertyWithDefaultValue.kt
@@ -17,7 +17,7 @@ import com.yandex.div.data.*
 
 @Mockable
 class EntityWithPropertyWithDefaultValue(
-    @JvmField final val int: Expression<Int> = INT_DEFAULT_VALUE, // constraint: number >= 0; default value: 0
+    @JvmField final val int: Expression<Long> = INT_DEFAULT_VALUE, // constraint: number >= 0; default value: 0
     @JvmField final val nested: Nested? = null,
     @JvmField final val url: Expression<Uri> = URL_DEFAULT_VALUE, // valid schemes: [https]; default value: https://yandex.ru
 ) : JSONSerializable {
@@ -34,7 +34,7 @@ class EntityWithPropertyWithDefaultValue(
     companion object {
         const val TYPE = "entity_with_property_with_default_value"
 
-        private val INT_DEFAULT_VALUE = Expression.constant(0)
+        private val INT_DEFAULT_VALUE = Expression.constant(0L)
         private val URL_DEFAULT_VALUE = Expression.constant(Uri.parse("https://yandex.ru"))
 
         @JvmStatic
@@ -48,8 +48,8 @@ class EntityWithPropertyWithDefaultValue(
             )
         }
 
-        private val INT_TEMPLATE_VALIDATOR = ValueValidator<Int> { it: Int -> it >= 0 }
-        private val INT_VALIDATOR = ValueValidator<Int> { it: Int -> it >= 0 }
+        private val INT_TEMPLATE_VALIDATOR = ValueValidator<Long> { it: Long -> it >= 0 }
+        private val INT_VALIDATOR = ValueValidator<Long> { it: Long -> it >= 0 }
         private val URL_TEMPLATE_VALIDATOR = ValueValidator<Uri> { it.hasScheme(listOf("https")) }
         private val URL_VALIDATOR = ValueValidator<Uri> { it.hasScheme(listOf("https")) }
 
@@ -59,7 +59,7 @@ class EntityWithPropertyWithDefaultValue(
 
     @Mockable
     class Nested(
-        @JvmField final val int: Expression<Int> = INT_DEFAULT_VALUE, // constraint: number >= 0; default value: 0
+        @JvmField final val int: Expression<Long> = INT_DEFAULT_VALUE, // constraint: number >= 0; default value: 0
         @JvmField final val nonOptional: Expression<String>,
         @JvmField final val url: Expression<Uri> = URL_DEFAULT_VALUE, // valid schemes: [https]; default value: https://yandex.ru
     ) : JSONSerializable {
@@ -73,7 +73,7 @@ class EntityWithPropertyWithDefaultValue(
         }
 
         companion object {
-            private val INT_DEFAULT_VALUE = Expression.constant(0)
+            private val INT_DEFAULT_VALUE = Expression.constant(0L)
             private val URL_DEFAULT_VALUE = Expression.constant(Uri.parse("https://yandex.ru"))
 
             @JvmStatic
@@ -87,8 +87,8 @@ class EntityWithPropertyWithDefaultValue(
                 )
             }
 
-            private val INT_TEMPLATE_VALIDATOR = ValueValidator<Int> { it: Int -> it >= 0 }
-            private val INT_VALIDATOR = ValueValidator<Int> { it: Int -> it >= 0 }
+            private val INT_TEMPLATE_VALIDATOR = ValueValidator<Long> { it: Long -> it >= 0 }
+            private val INT_VALIDATOR = ValueValidator<Long> { it: Long -> it >= 0 }
             private val URL_TEMPLATE_VALIDATOR = ValueValidator<Uri> { it.hasScheme(listOf("https")) }
             private val URL_VALIDATOR = ValueValidator<Uri> { it.hasScheme(listOf("https")) }
