+++ client/android/div-data/src/main/java/com/yandex/div/util/DivDataUtils.kt
@@ -2,7 +2,7 @@ package com.yandex.div.util
 
 import com.yandex.div2.DivData
 
-val DivData.Companion.INVALID_STATE_ID get() = -1
+val DivData.Companion.INVALID_STATE_ID get() = -1L
 
-fun DivData.getInitialStateId(): Int =
+fun DivData.getInitialStateId(): Long =
     if (states.isEmpty()) DivData.INVALID_STATE_ID else states[0].stateId
